refactor: extract prolly_btree_cursor_payload.c
Move IntegerKey and payload access APIs into a dedicated TU. Share
tree-value and int-key helpers as static inlines in prolly_btree_int.h
without rewriting them to Span. Promote materializeDeferredMergedSeekBackward
for payload entry points. Wire build, amalgamation, and lint.
